feat: Add image preview functionality to FormMarca component
The code changes include adding image preview functionality to the FormMarca component. Users can now select an image file for the brand logo, and a preview of the selected image will be displayed. This enhances the user experience by allowing them to see a visual representation of the logo before submitting the form.